needed to add the additional logger gunicorn configuration in order for metadata to work as expected
